Video Autoplay Feature addition script.
This script was a minor task to get around slow moving function of a website video playback. Since the human interaction it took to click each video added more time to the training that consisted of 100+ hours of video I built this script to automate the flow of the videos so I could take better notes and consistantly flow through the content.